+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Improved
 
 * Ensure all valves are closed at the end of the LN2 fill process.
+* Do not run post-fill validation if the fill was aborted or failed.
 
 
 ## 0.3.18 - 2025-12-24