+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Ensure all valves are closed at the end of the LN2 fill process.
 * Do not run post-fill validation if the fill was aborted or failed.
+* Add retrier for `LN2Handler.stop()`.
 
 
 ## 0.3.18 - 2025-12-24